+++ lib/modules/test/test_service.dart
@@ -1,9 +1,9 @@
 ﻿// Module: test
-// Updated at: 2026-07-11 22:29:54
-// Commit: feat(test): add bi-directional event bus handler for test
+// Updated at: 2026-07-12 15:04:15
+// Commit: perf(test): optimize memory consumption in test list viewport
 
 class testComponentState {
-  final String commitRef = '2721';
-  final String lastMessage = 'feat(test): add bi-directional event bus handler for test';
-  final DateTime timestamp = DateTime.parse('2026-07-11T22:29:54+05:00');
+  final String commitRef = '2748';
+  final String lastMessage = 'perf(test): optimize memory consumption in test list viewport';
+  final DateTime timestamp = DateTime.parse('2026-07-12T15:04:15+05:00');
 }